Add 2 java classes to be used for the GUI
User and Entertainment are the 2 java classes the GUI will call on. They return a single user/entertainment based on the ids or all the available info. These classes provide methods to either add an object or get an object from the database. They also have methods to change and get all of the information for each. There are three custom exceptions just to ensure the objects were added to the database properly and if the entertainment object was retrieved properly. the Connect class is how we connect to the database. I don't think the GUI should call on this method, but its there just in case. I think we should only have one class that stores our database connection info.