Comments, Transactions, and Fixes
Added comments to all files to make things easier to find and understand what they are doing. Added all of the needed transactions. I tested them, but they should be tested again. Found a couple of errors with some of the queries, especially in Entertainment with getting the sequals and some of the searches